+++ blocks/site-stats/index.asset.php
@@ -1,0 +1,14 @@
+<?php
+
+defined('ABSPATH') || exit;
+
+return array(
+    'dependencies' => array(
+        'wp-block-editor',
+        'wp-blocks',
+        'wp-components',
+        'wp-element',
+        'wp-i18n',
+    ),
+    'version' => '3.2.0',
+);